requirements.txt
websocket-client
requests
numpy
matplotlib
